move devices to p_roc_devices
